@@ -8,11 +8,35 @@
       "type": "package",
       "package": {
         "name": "medialize/uri.js",
-        "version": "1.18.2",
+        "version": "1.19.0",
         "source": {
           "url": "https://github.com/medialize/URI.js.git",
           "type": "git",
-          "reference": "v1.18.2"
+          "reference": "v1.19.0"
+        }
+      }
+    },
+    "lscache": {
+      "type": "package",
+      "package": {
+        "name": "pamelafox/lscache",
+        "version": "1.0.5",
+        "source": {
+          "url": "https://github.com/pamelafox/lscache.git",
+          "type": "git",
+          "reference": "1.0.5"
+        }
+      }
+    },
+    "malihu-custom-scrollbar": {
+      "type": "package",
+      "package": {
+        "name": "malihu/malihu-custom-scrollbar-plugin",
+        "version": "3.1.5",
+        "source": {
+          "url": "https://github.com/malihu/malihu-custom-scrollbar-plugin.git",
+          "type": "git",
+          "reference": "3.1.5"
         }
       }
     },
@@ -30,30 +54,36 @@
     } 
   },
   "require": {
-    "components/jquery": "1.12.*",
+    "components/jquery": "2.2.*",
     "components/jqueryui": "1.12.*",
-    "components/handlebars.js": "v1.3.0",
+    "components/handlebars.js": "v4.7.6",
     "davidstutz/bootstrap-multiselect": "v0.9.13",
     "easyrdf/easyrdf": "0.10.0-alpha.1",
-    "twig/twig": "1.29.*",
-    "twig/extensions": "1.4.*",
-    "twitter/bootstrap": "3.2.*",
+    "etdsolutions/waypoints": "4.0.0",
+    "twig/twig": "2.13.*",
+    "twig/extensions": "1.5.*",
+    "twitter/bootstrap": "3.3.*",
     "twitter/typeahead.js": "v0.10.5",
-    "willdurand/negotiation": "1.5.*",
+    "willdurand/negotiation": "2.3.*",
     "vakata/jstree": "3.3.*",
-    "punic/punic": "1.6.0",
+    "punic/punic": "3.5.1",
     "ml/json-ld": "1.*",
-    "medialize/uri.js": "1.18.2",
+    "medialize/uri.js": "1.19.0",
     "ext-gettext": "*",
-    "monolog/monolog": "^1.22"
+    "ext-intl": "*",
+    "ext-mbstring": "*",
+    "ext-xsl": "*",
+    "monolog/monolog": "1.23.*",
+    "newerton/jquery-mousewheel": "dev-master",
+    "pamelafox/lscache": "1.0.5",
+    "malihu/malihu-custom-scrollbar-plugin": "3.1.5",
+    "grimmlink/qtip2": "3.0.3"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.8 - 5.1",
-    "umpirsky/twig-gettext-extractor": "1.1.*",
-    "phpdocumentor/phpdocumentor": "2.*",
-    "codeclimate/php-test-reporter": "0.3.2",
-    "symfony/dom-crawler": "3.2.6",
-    "mockery/mockery": "1.0.0-alpha1"
+    "phpunit/phpunit": "7.2.0 - 7.5.20",
+    "umpirsky/twig-gettext-extractor": "1.3.*",
+    "symfony/dom-crawler": "3.4.3",
+    "mockery/mockery": "1.3.1"
   },
   "autoload": {
     "classmap": ["controller/", "model/", "model/sparql/"]
