@@ -1,19 +1,20 @@
 <?php
 
-class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
+class ConceptPropertyValueTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  private $model;
   private $concept;
   private $vocab;
-    
+
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('test');
     $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta112', 'en');
     $this->concept = reset($results);
@@ -24,6 +25,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
    */
   public function testConstructor() {
     $mockres = $this->getMockBuilder('EasyRdf\\Resource')->disableOriginalConstructor()->getMock();
+    $mockres->method('getUri')->willReturn('http://example.org/');
     $propval = new ConceptPropertyValue($this->model, $this->vocab, $mockres, 'skosmos:testProp', 'en');
     $this->assertInstanceOf('ConceptPropertyValue', $propval);
   }
@@ -34,7 +36,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
   public function testGetLabel() {
     $props = $this->concept->getProperties();
     $propvals = $props['skos:narrower']->getValues();
-    $this->assertEquals('Crucian carp', $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getLabel());
+    $this->assertEquals('Crucian carp', $propvals['Crucian carp http://www.skosmos.skos/test/ta121']->getLabel());
   }
 
   /**
@@ -57,25 +59,25 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('english lit', $mapping->getLabel('en'));
     $this->assertEquals('default lit', $mapping->getLabel());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::getType
    */
   public function testGetType() {
     $props = $this->concept->getProperties();
     $propvals = $props['skos:narrower']->getValues();
-    $this->assertEquals('skos:narrower', $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getType());
+    $this->assertEquals('skos:narrower', $propvals['Crucian carp http://www.skosmos.skos/test/ta121']->getType());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::getUri
    */
   public function testGetUri() {
     $props = $this->concept->getProperties();
     $propvals = $props['skos:narrower']->getValues();
-    $this->assertEquals('http://www.skosmos.skos/test/ta121', $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getUri());
+    $this->assertEquals('http://www.skosmos.skos/test/ta121', $propvals['Crucian carp http://www.skosmos.skos/test/ta121']->getUri());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::getVocab
    */
@@ -83,7 +85,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $props = $this->concept->getProperties();
     $vocab = $this->model->getVocabulary('test');
     $propvals = $props['skos:narrower']->getValues();
-    $this->assertEquals($vocab, $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getVocab());
+    $this->assertEquals($vocab, $propvals['Crucian carp http://www.skosmos.skos/test/ta121']->getVocab());
   }
 
   /**
@@ -92,7 +94,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
   public function testGetVocabName() {
     $props = $this->concept->getProperties();
     $propvals = $props['skos:narrower']->getValues();
-    $this->assertEquals('Test ontology', $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getVocabName());
+    $this->assertEquals('Test short', $propvals['Crucian carp http://www.skosmos.skos/test/ta121']->getVocabName());
   }
 
   /**
@@ -103,7 +105,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $concept = reset($results);
     $props = $concept->getProperties();
     $propvals = $props['skos:broader']->getValues();
-    $this->assertEquals(665, $propvals['665Carphttp://www.skosmos.skos/test/ta112']->getNotation());
+    $this->assertEquals(665, $propvals['665 Carp http://www.skosmos.skos/test/ta112']->getNotation());
   }
 
   /**
@@ -111,10 +113,11 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
    */
   public function testGetNotationWhenThereIsNone() {
     $mockres = $this->getMockBuilder('EasyRdf\\Resource')->disableOriginalConstructor()->getMock();
+    $mockres->method('getUri')->willReturn('http://example.org/');
     $propval = new ConceptPropertyValue($this->model, $this->vocab, $mockres, 'en');
     $this->assertEquals(null, $propval->getNotation());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::__toString
    */
@@ -123,9 +126,9 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $concept = reset($results);
     $props = $concept->getProperties();
     $propvals = $props['skos:broader']->getValues();
-    $this->assertEquals('Carp', (string)$propvals['665Carphttp://www.skosmos.skos/test/ta112']);
+    $this->assertEquals('Carp', (string)$propvals['665 Carp http://www.skosmos.skos/test/ta112']);
   }
-  
+
   /**
    * @covers ConceptPropertyValue::__toString
    */
@@ -142,9 +145,9 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $mockres->method('get')->will($this->returnValue($mocklit));
     $mockres->method('getUri')->will($this->returnValue('http://thisdoesntexistatalland.sefsf/2j2h4/'));
     $propval = new ConceptPropertyValue($this->model, $mockvoc, $mockres, null);
-    $this->assertEquals('T3STTerm label', (string)$propval);
+    $this->assertEquals('T3ST Term label', (string)$propval);
   }
-   
+
   /**
    * @covers ConceptPropertyValue::__toString
    */
@@ -198,7 +201,18 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
    */
   public function testGetSubMembersEmpty() {
     $mockres = $this->getMockBuilder('EasyRdf\\Resource')->disableOriginalConstructor()->getMock();
+    $mockres->method('getUri')->willReturn('http://example.org/');
     $propval = new ConceptPropertyValue($this->model, $this->vocab, $mockres, 'en');
     $this->assertEquals(null, $propval->getSubMembers());
+  }
+
+  public function testGetReifiedPropertyValues() {
+    $vocab = $this->model->getVocabulary('xl');
+    $concept = $vocab->getConceptInfo('http://www.skosmos.skos/xl/c1', 'en')[0];
+    $props = $concept->getProperties();
+    $vals = $props['skos:definition']->getValues();
+    $val = reset($vals);
+    $reified_vals = $val->getReifiedPropertyValues();
+    $this->assertEquals(2, count($reified_vals));
   }
 }
