@@ -1,21 +1,32 @@
 <?php
 
-class ConceptTest extends PHPUnit_Framework_TestCase
+class ConceptTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  /**
+   * @var Model
+   */
+  private $model;
   private $concept;
+  private $cbdVocab;
+  private $cbdGraph;
 
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('test');
     $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta112', 'en');
     $this->concept = reset($results);
+
+    $this->cbdVocab = $this->model->getVocabulary('cbd');
+    $this->cbdGraph =  new EasyRdf\Graph();
+    $this->cbdGraph->parseFile(__DIR__ . "/test-vocab-data/cbd.ttl", "turtle");
+
   }
 
   /**
@@ -28,7 +39,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $this->assertInstanceOf('Concept', $concept);
     $this->assertEquals('Test ontology', $concept->getVocabTitle());
   }
-  
+
   /**
    * @covers Concept::getUri
    */
@@ -37,7 +48,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $uri = $this->concept->getURI();
     $this->assertEquals('http://www.skosmos.skos/test/ta112', $uri);
   }
-  
+
   /**
    * @covers Concept::getDeprecated
    */
@@ -46,7 +57,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $deprecated = $this->concept->getDeprecated();
     $this->assertEquals(false, $deprecated);
   }
-  
+
   /**
    * @covers Concept::getVocab
    */
@@ -55,7 +66,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $voc = $this->concept->getVocab();
     $this->assertInstanceOf('Vocabulary', $voc);
   }
-  
+
   /**
    * @covers Concept::getVocabTitle
    */
@@ -64,7 +75,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $title = $this->concept->getVocabTitle();
     $this->assertEquals('Test ontology', $title);
   }
-  
+
   /**
    * @covers Concept::getShortName
    */
@@ -82,7 +93,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $fb = $this->concept->getFoundBy();
     $this->assertEquals(null, $fb);
   }
-  
+
   /**
    * @covers Concept::setFoundBy
    * @covers Concept::getFoundByType
@@ -97,18 +108,38 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('testing matched label', $fb);
     $this->assertEquals('alt', $fbtype);
   }
-  
+
   /**
    * @covers Concept::getForeignLabels
-   * @covers Concept::literalLanguageToString
+   * @covers Concept::getForeignLabelList
+   * @covers Concept::langToString
    */
   public function testGetForeignLabels()
   {
     $labels = $this->concept->getForeignLabels();
 
-    $this->assertEquals('Karppi', $labels['Finnish'][0]->getLabel());
+    $this->assertEquals('Karppi', $labels['Finnish']['prefLabel'][0]->getLabel());
+    $this->assertArrayNotHasKey('altLabel', $labels['Finnish']);
+    $this->assertArrayNotHasKey('English', $labels);
+
+    $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta115', 'en');
+    $concept = reset($results);
+    $this->assertEmpty($concept->getForeignLabels());
+
+    $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta127', 'en');
+    $concept = reset($results);
+    $labels = $concept->getForeignLabels();
+    $this->assertEquals(['', 'Finnish', 'Swedish'], array_keys($labels));
+
+    $this->assertEquals("Ä before 'A first' in English (default) collation", $labels['']['prefLabel'][0]->getLabel());
+    $this->assertEquals("Test sorting labels 2", $labels['']['prefLabel'][3]->getLabel());
+
+    $this->assertArrayHasKey('prefLabel', $labels['Finnish']);
+    $fiAltLabels = array_map(function ($elem) {return $elem->getLabel();}, $labels['Finnish']['altLabel']);
+    $fiCorrectSort = ["A sort first", "B sorts second", "Ä way after B in Finnish collation"];
+    $this->assertEquals($fiCorrectSort, $fiAltLabels);
   }
-  
+
   /**
    * @covers Concept::getAllLabels
    */
@@ -120,7 +151,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('Iljettävä limanuljaska', $labels['Finnish'][0]->getLabel());
     $this->assertEquals('any fish belonging to the order Anguilliformes', $labels['English'][0]->getLabel());
   }
-  
+
   /**
    * @covers Concept::getProperties
    * @covers ConceptProperty::getValues
@@ -143,7 +174,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
   public function testGetPropertiesCorrectNumberOfProperties()
   {
     $props = $this->concept->getProperties();
- 
+
     $this->assertEquals(6, sizeof($props));
   }
 
@@ -159,9 +190,9 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $props = $this->concept->getProperties();
     $expected = array (0 => 'rdf:type', 1=> 'skos:broader',2 => 'skos:narrower',3 => 'skos:altLabel',4 => 'skos:scopeNote',5 => 'http://www.skosmos.skos/testprop');
     $this->assertEquals($expected, array_keys($props));
- 
+
   }
-  
+
   /**
    * @covers Concept::getProperties
    * @covers ConceptProperty::getValues
@@ -169,18 +200,18 @@ class ConceptTest extends PHPUnit_Framework_TestCase
    */
   public function testGetPropertiesAlphabeticalSortingOfPropertyValues()
   {
-    $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta1', 'en'); 
+    $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta1', 'en');
     $concept = reset($results);
     $props = $concept->getProperties();
-    $prevlabel;
-    foreach($props['skos:narrower'] as $val) {
-      $label = is_string($val->getLabel()) ? $val->getLabel() : $val->getLabel()-getValue();
+    $prevlabel = null;
+    foreach($props['skos:narrower']->getValues() as $val) {
+      $label = $val->getLabel();
       if ($prevlabel)
-        $this->assertEquals(1, strnatcmp($prevlabel, $label));
+        $this->assertEquals(-1, strnatcasecmp($prevlabel, $label));
       $prevlabel = $label;
     }
   }
-  
+
   /**
    * @covers Concept::getMappingProperties
    * @covers ConceptProperty::getValues
@@ -193,7 +224,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $values = $props['skos:closeMatch']->getValues();
     $this->assertCount(2, $values);
   }
-  
+
   /**
    * @covers Concept::removeDuplicatePropertyValues
    * @covers Concept::getProperties
@@ -205,7 +236,31 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $props = $concept->getProperties();
     $this->assertCount(1, $props);
   }
-  
+
+  /**
+   * @covers Concept::removeDuplicatePropertyValues
+   * @covers Concept::getPreferredSubpropertyLabelTranslation
+   */
+  public function testgetPreferredSubpropertyLabelTranslation() {
+    $vocab = $this->model->getVocabulary('duplicates');
+    $concepts = $vocab->getConceptInfo("http://www.skosmos.skos/dup/d6", "en");
+    $concept = $concepts[0];
+    $this->assertEquals($concept->getPreferredSubpropertyLabelTranslation('en'), "Subproperty of skos:prefLabel");
+    $this->assertEquals($concept->getPreferredSubpropertyLabelTranslation('fi'), null);
+  }
+
+  /**
+   * @covers Concept::removeDuplicatePropertyValues
+   * @covers Concept::getProperties
+   */
+  public function testRemoveDuplicateValuesForPreflabel() {
+    $vocab = $this->model->getVocabulary('duplicates');
+    $concepts = $vocab->getConceptInfo("http://www.skosmos.skos/dup/d7", "en");
+    $concept = $concepts[0];
+    $props = $concept->getProperties();
+    $this->assertCount(0, $props);
+  }
+
   /**
    * @covers Concept::removeDuplicatePropertyValues
    * @covers Concept::getProperties
@@ -247,10 +302,9 @@ class ConceptTest extends PHPUnit_Framework_TestCase
    * @covers Concept::getDate
    * @covers ConceptProperty::getValues
    * @covers ConceptPropertyValueLiteral::getLabel
-   * @expectedException PHPUnit_Framework_Error
    */
-
   public function testGetTimestampInvalidWarning() {
+    $this->expectException(PHPUnit\Framework\Error\Error::class);
     $vocab = $this->model->getVocabulary('test');
     $concepts = $vocab->getConceptInfo("http://www.skosmos.skos/test/ta114", "en");
     $concept = $concepts[0];
@@ -280,10 +334,10 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $props = $this->concept->getProperties();
     $propvals = $props['rdf:type']->getValues();
     $this->assertCount(1, $propvals); // should only have type meta:TestClass, not skos:Concept (see #200)
-    $this->assertEquals('Test class', $propvals['Test classhttp://www.skosmos.skos/test-meta/TestClass']->getLabel());
-    $this->assertEquals('http://www.skosmos.skos/test-meta/TestClass', $propvals['Test classhttp://www.skosmos.skos/test-meta/TestClass']->getUri());
+    $this->assertEquals('Test class', $propvals['Test class http://www.skosmos.skos/test-meta/TestClass']->getLabel());
+    $this->assertEquals('http://www.skosmos.skos/test-meta/TestClass', $propvals['Test class http://www.skosmos.skos/test-meta/TestClass']->getUri());
   }
-  
+
   /**
    * @covers Concept::getNotation
    */
@@ -294,7 +348,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $concept = $concepts[0];
     $this->assertEquals(null, $concept->getNotation());
   }
-  
+
   /**
    * @covers Concept::getNotation
    */
@@ -302,7 +356,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
   {
     $this->assertEquals('665', $this->concept->getNotation());
   }
-  
+
   /**
    * @covers Concept::getLabel
    */
@@ -310,18 +364,18 @@ class ConceptTest extends PHPUnit_Framework_TestCase
   {
     $this->assertEquals('Carp', $this->concept->getLabel()->getValue());
   }
-  
+
   /**
    * @covers Concept::getLabel
    */
   public function testGetLabelWhenNull()
   {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $vocab = $model->getVocabulary('test');
     $concept = $vocab->getConceptInfo("http://www.skosmos.skos/test/ta120", "en");
     $this->assertEquals(null, $concept[0]->getLabel());
   }
-  
+
   /**
    * @covers Concept::getLabel
    * @covers Concept::setContentLang
@@ -333,15 +387,15 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('pl', $this->concept->getContentLang());
     $this->assertEquals('Carp', $this->concept->getLabel()->getValue());
   }
-  
+
   /**
    * @covers Concept::getArrayProperties
    * @covers Concept::getGroupProperties
-   * @covers Concept::getReverseResources
+   * @covers Concept::getCollections
    */
   public function testGetGroupProperties()
   {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $vocab = $model->getVocabulary('groups');
     $concept = $vocab->getConceptInfo("http://www.skosmos.skos/groups/ta111", "en");
     $arrays = $concept[0]->getArrayProperties();
@@ -350,18 +404,33 @@ class ConceptTest extends PHPUnit_Framework_TestCase
     $groups = $concept[0]->getGroupProperties();
     $this->assertEmpty($groups);
   }
-  
+
   /**
    * @covers Concept::getGroupProperties
-   * @covers Concept::getReverseResources
+   * @covers Concept::getCollections
    */
   public function testGetGroupPropertiesWithDuplicatedInformationFilteredOut()
   {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $vocab = $model->getVocabulary('dupgroup');
     $concept = $vocab->getConceptInfo("http://www.skosmos.skos/dupgroup/c1", "en");
     $groups = $concept[0]->getGroupProperties();
     $this->assertEquals(0, sizeof($groups));
+  }
+
+  /**
+   * @covers Concept::getGroupProperties
+   * @covers Concept::getCollections
+   */
+  public function testGetGroupPropertiesWithHierarchy()
+  {
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
+    $vocab = $model->getVocabulary('dupgroup');
+    $concept = $vocab->getConceptInfo("http://www.skosmos.skos/dupgroup/ta111", "en");
+    $groups = $concept[0]->getGroupProperties();
+    $this->assertEquals(2, sizeof($groups));
+    $this->assertArrayHasKey("Animalia", $groups);
+    $this->assertArrayHasKey("Biology", $groups);
   }
 
   /**
@@ -373,7 +442,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
    */
   public function testGetPropertiesWithNarrowersPartOfACollection()
   {
-    $model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $vocab = $model->getVocabulary('groups');
     $concept = $vocab->getConceptInfo("http://www.skosmos.skos/groups/ta1", "en");
     $props = $concept[0]->getProperties();
@@ -391,7 +460,7 @@ class ConceptTest extends PHPUnit_Framework_TestCase
       }
     }
   }
-  
+
   /**
    * @covers Concept::getProperties
    */
@@ -408,13 +477,92 @@ class ConceptTest extends PHPUnit_Framework_TestCase
    * @covers Concept::getProperties
    * @covers ConceptProperty::getValues
    */
-  
   public function testGetPropertiesDefinitionResource() {
     $vocab = $this->model->getVocabulary('test');
     $concepts = $vocab->getConceptInfo('http://www.skosmos.skos/test/ta122', 'en');
     $concept = $concepts[0];
     $props = $concept->getProperties();
     $propvals = $props['skos:definition']->getValues();
-    $this->assertEquals('The black sea bass (Centropristis striata) is an exclusively marine fish.', $propvals['The black sea bass (Centropristis striata) is an exclusively marine fish.http://www.skosmos.skos/test/black_sea_bass_def']->getLabel());
+    $this->assertEquals('The black sea bass (Centropristis striata) is an exclusively marine fish.', $propvals['The black sea bass (Centropristis striata) is an exclusively marine fish. http://www.skosmos.skos/test/black_sea_bass_def']->getLabel());
+  }
+
+  /**
+   * @covers Concept::addResourceReifications
+   * @covers Concept::addLiteralReifications
+   */
+  public function testExternalResourceReifications() {
+    $concepts = $this->cbdVocab->getConceptInfo('http://www.skosmos.skos/cbd/test2', 'en');
+    $concept = $concepts[0];
+
+    $res = $this->cbdGraph->resource('http://www.skosmos.skos/cbd/test1');
+
+    $concept->processExternalResource($res);
+    $json =  $concept->dumpJsonLd();
+    $error_count = substr_count($json, "REIFICATION_ERROR");
+    $this->assertEquals($error_count, 0);
+  }
+
+  /**
+   * @covers Concept::processExternalResource
+   * @covers Concept::addExternalTriplesToGraph
+   * @covers Concept::addPropertyValues
+   */
+  public function testProcessExternalResource() {
+    $concepts = $this->cbdVocab->getConceptInfo('http://www.skosmos.skos/cbd/test2', 'en');
+    $concept = $concepts[0];
+
+    $res = $this->cbdGraph->resource('http://www.skosmos.skos/cbd/test1');
+
+    $concept->processExternalResource($res);
+    $json =  $concept->dumpJsonLd();
+    $this->assertContains('HY', $json);
+    $this->assertContains('AK', $json);
+    $this->assertContains('OS', $json);
+    $contains_count = substr_count($json, "CONTAINS");
+    $this->assertEquals($contains_count, 3);
+  }
+
+  /**
+   * Data provider for testGetModifiedDate test method.
+   * @return array
+   */
+  public function modifiedDateDataProvider() {
+    return [
+      ["cat", "2018-12-13T06:28:14", "+00:00"],  # set #0
+      ["dog", "2018-12-13T06:28:14", "+00:00"],  # set #1
+      ["owl", "2018-10-22T00:00:00", "+00:00"],  # set #2
+      ["parrot", "2018-10-22T00:00:00", "+00:00"],  # set #3
+      ["macaw", "2018-10-22T12:34:45", "+00:00"],  # set #4
+      ["sloth", "2018-10-22T12:34:45", "+05:30"],  # set #5
+    ];
+  }
+
+  /**
+   * @covers Concept::getModifiedDate
+   * @dataProvider modifiedDateDataProvider
+   * @throws Exception if it fails to load the vocabulary
+   */
+  public function testGetModifiedDate($animal, $expected_time, $expected_timezone) {
+    $vocab = $this->model->getVocabulary('http304');
+    $results = $vocab->getConceptInfo('http://www.skosmos.skos/test/' . $animal, 'en');
+    $concept = reset($results);
+    if (is_null($expected_time)) {
+        $modifiedDate = $concept->getModifiedDate();
+        $this->assertNull($modifiedDate);
+    } else {
+        $modifiedDate = $concept->getModifiedDate();
+        $this->assertEquals($expected_time . $expected_timezone, $modifiedDate->format("c"));
+    }
+  }
+
+  /**
+   * @covers Concept::getModifiedDate
+   */
+  public function testGetModifiedDateFallbackToVocabularyModified() {
+    $vocab = $this->model->getVocabulary('test');
+    $results = $vocab->getConceptInfo('http://www.skosmos.skos/test/ta111', 'en');
+    $concept = reset($results);
+    $modifiedDate = $concept->getModifiedDate();
+    $this->assertEquals(new DateTime("2014-10-01T16:29:03+00:00"), $modifiedDate);
   }
 }
