@@ -1,36 +1,37 @@
 <?php
 
-class GenericSparqlTest extends PHPUnit_Framework_TestCase
+class GenericSparqlTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
-  private $graph; 
+  private $model;
+  private $graph;
   private $sparql;
   private $vocab;
   private $params;
 
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('test');
     $this->graph = $this->vocab->getGraph();
     $this->params = $this->getMockBuilder('ConceptSearchParameters')->disableOriginalConstructor()->getMock();
-    $this->sparql = new GenericSparql('http://localhost:3030/ds/sparql', $this->graph, $this->model);
+    $this->sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $this->graph, $this->model);
   }
- 
+
   /**
    * @covers GenericSparql::__construct
    */
   public function testConstructor() {
-    $gs = new GenericSparql('http://localhost:3030/ds/sparql', $this->graph, $this->model);
+    $gs = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $this->graph, $this->model);
     $this->assertInstanceOf('GenericSparql', $gs);
   }
-  
+
   /**
    * @covers GenericSparql::getGraph
    */
   public function testGetGraph() {
-    $gs = new GenericSparql('http://localhost:3030/ds/sparql', $this->graph, $this->model);
+    $gs = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $this->graph, $this->model);
     $this->assertEquals($this->graph, $gs->getGraph());
   }
 
@@ -41,9 +42,9 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    */
   public function testCountConcepts() {
     $actual = $this->sparql->countConcepts();
-    $this->assertEquals(16, $actual['http://www.w3.org/2004/02/skos/core#Concept']['count']);
+    $this->assertEquals(17, $actual['http://www.w3.org/2004/02/skos/core#Concept']['count']);
   }
-  
+
   /**
    * @covers GenericSparql::countLangConcepts
    * @covers GenericSparql::generateCountLangConceptsQuery
@@ -54,7 +55,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(11, $actual['en']['skos:prefLabel']);
     $this->assertEquals(1, $actual['en']['skos:altLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::countLangConcepts
    * @covers GenericSparql::generateCountLangConceptsQuery
@@ -64,9 +65,9 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->countLangConcepts(array('en','fi'));
     $this->assertEquals(11, $actual['en']['skos:prefLabel']);
     $this->assertEquals(1, $actual['en']['skos:altLabel']);
-    $this->assertEquals(2, $actual['fi']['skos:prefLabel']);
+    $this->assertEquals(3, $actual['fi']['skos:prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryFirstCharacters
    * @covers GenericSparql::generateFirstCharactersQuery
@@ -85,7 +86,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->queryLabel('http://notfound', null);
     $this->assertEquals(null, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryLabel
    */
@@ -93,7 +94,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->queryLabel('http://www.skosmos.skos/test/ta120', null);
     $this->assertEquals(array(), $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -124,7 +125,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -150,7 +151,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -176,7 +177,134 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
+  /**
+   * @covers GenericSparql::queryConceptsAlphabetical
+   * @covers GenericSparql::generateAlphabeticalListQuery
+   * @covers GenericSparql::transformAlphabeticalListResults
+   */
+  public function testQualifiedNotationAlphabeticalList() {
+    $voc = $this->model->getVocabulary('test-qualified-notation');
+    $res = new EasyRdf\Resource("http://www.w3.org/2004/02/skos/core#notation");
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $voc->getGraph(), $this->model);
+
+    $actual = $sparql->queryConceptsAlphabetical("a", "en", null, null, null, false, $res);
+
+    $expected = array (
+      0 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn1',
+        'localname' => 'qn1',
+        'prefLabel' => 'A',
+        'lang' => 'en',
+        'qualifier' => 'A'
+      ),
+      1 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn1b',
+        'localname' => 'qn1b',
+        'prefLabel' => 'A',
+        'lang' => 'en',
+        'qualifier' => 'A'
+      ),
+      2 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn1c',
+        'localname' => 'qn1c',
+        'prefLabel' => 'A',
+        'lang' => 'en',
+      ),
+    );
+    $this->assertEquals($expected, $actual);
+
+    $actual = $sparql->queryConceptsAlphabetical("b", "en", null, null, null, false, $res);
+
+    $expected = array (
+      0 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn2',
+        'localname' => 'qn2',
+        'prefLabel' => 'B',
+        'lang' => 'en',
+        'qualifier' => 'B'
+      ),
+      1 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn2',
+        'localname' => 'qn2',
+        'prefLabel' => 'B',
+        'lang' => 'en',
+        'qualifier' => 'C'
+      ),
+      2 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn2b',
+        'localname' => 'qn2b',
+        'prefLabel' => 'B',
+        'lang' => 'en',
+        'qualifier' => 'B'
+      ),
+      3 => array (
+        'uri' => 'http://www.skosmos.skos/test/qn2b',
+        'localname' => 'qn2b',
+        'prefLabel' => 'B',
+        'lang' => 'en',
+        'qualifier' => 'C'
+      ),
+    );
+    $this->assertEquals($expected, $actual);
+  }
+
+  /**
+   * @covers GenericSparql::queryConceptsAlphabetical
+   * @covers GenericSparql::generateAlphabeticalListQuery
+   * @covers GenericSparql::transformAlphabeticalListResults
+   */
+  public function testQualifiedBroaderAlphabeticalList() {
+    $voc = $this->model->getVocabulary('test-qualified-broader');
+    $res = new EasyRdf\Resource("http://www.w3.org/2004/02/skos/core#broader");
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $voc->getGraph(), $this->model);
+
+    $actual = $sparql->queryConceptsAlphabetical("a", "en", null, null, null, false, $res);
+
+    $expected = array (
+      0 => array (
+        'uri' => 'http://www.skosmos.skos/test/qb1',
+        'localname' => 'qb1',
+        'prefLabel' => 'A',
+        'lang' => 'en',
+      ),
+    );
+    $this->assertEquals($expected, $actual);
+
+    $actual = $sparql->queryConceptsAlphabetical("b", "en", null, null, null, false, $res);
+
+    $expected = array (
+      0 => array (
+        'uri' => 'http://www.skosmos.skos/test/qb2',
+        'localname' => 'qb2',
+        'prefLabel' => 'B',
+        'lang' => 'en',
+        'qualifier' => 'qb1'
+      ),
+    );
+    $this->assertEquals($expected, $actual);
+
+    $actual = $sparql->queryConceptsAlphabetical("c", "en", null, null, null, false, $res);
+
+    $expected = array (
+      0 => array (
+        'uri' => 'http://www.skosmos.skos/test/qb3',
+        'localname' => 'qb3',
+        'prefLabel' => 'C',
+        'lang' => 'en',
+        'qualifier' => 'qb1'
+      ),
+      1 => array (
+        'uri' => 'http://www.skosmos.skos/test/qb3',
+        'localname' => 'qb3',
+        'prefLabel' => 'C',
+        'lang' => 'en',
+        'qualifier' => 'qb2'
+      ),
+    );
+    $this->assertEquals($expected, $actual);
+  }
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -195,12 +323,26 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    * @covers GenericSparql::formatFilterConditions
    * @covers GenericSparql::transformAlphabeticalListResults
    */
+  public function testQueryConceptsAlphabeticalMatchLanguage() {
+    $actual = $this->sparql->queryConceptsAlphabetical('e', 'en');
+    // there are two prefLabels starting with E, "Eel"@en and "Europa"@nb
+    // checking that only the first one is returned
+    $this->assertEquals(1, sizeof($actual));
+    $this->assertEquals('Eel', $actual[0]['prefLabel']);
+  }
+
+  /**
+   * @covers GenericSparql::queryConceptsAlphabetical
+   * @covers GenericSparql::generateAlphabeticalListQuery
+   * @covers GenericSparql::formatFilterConditions
+   * @covers GenericSparql::transformAlphabeticalListResults
+   */
   public function testQueryConceptsAlphabeticalSpecialChars() {
     $actual = $this->sparql->queryConceptsAlphabetical('!*', 'en');
     $this->assertEquals(1, sizeof($actual));
     $this->assertEquals('-"special" character \\example\\', $actual[0]['prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -212,7 +354,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, sizeof($actual));
     $this->assertContains("3D", $actual[0]['prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -235,7 +377,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    */
   public function testQueryConceptInfoWithMultipleVocabs()
   {
-    $this->sparql = new GenericSparql('http://localhost:3030/ds/sparql', '?graph', $this->model);
+    $this->sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', '?graph', $this->model);
     $voc2 = $this->model->getVocabulary('test');
     $voc3 = $this->model->getVocabulary('dates');
     $voc4 = $this->model->getVocabulary('groups');
@@ -257,7 +399,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    */
   public function testQueryConceptInfoWithAllVocabs()
   {
-    $this->sparql = new GenericSparql('http://localhost:3030/ds/sparql', '?graph', $this->model);
+    $this->sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', '?graph', $this->model);
     $actual = $this->sparql->queryConceptInfo(array('http://www.skosmos.skos/test/ta121', 'http://www.skosmos.skos/groups/ta111'), null, null, 'en');
     $this->assertInstanceOf('Concept', $actual[0]);
     $this->assertEquals('http://www.skosmos.skos/test/ta121', $actual[0]->getUri());
@@ -334,7 +476,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $actual = $this->sparql->queryConceptScheme('http://www.skosmos.skos/test/conceptscheme');
     $this->assertInstanceOf('EasyRdf\Graph', $actual);
-    $this->assertEquals('http://localhost:3030/ds/sparql', $actual->getUri());
+    $this->assertEquals('http://localhost:13030/skosmos-test/sparql', $actual->getUri());
   }
 
   /**
@@ -349,6 +491,49 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       $this->assertEquals('http://www.skosmos.skos/test/conceptscheme', $scheme);
       $this->assertEquals('Test conceptscheme', $label['label']);
     }
+  }
+
+  /**
+   * @covers GenericSparql::queryConceptSchemes
+   * @covers GenericSparql::generateQueryConceptSchemesQuery
+   * @covers GenericSparql::transformQueryConceptSchemesResults
+   */
+  public function testQueryConceptSchemesSubject()
+  {
+      $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', 'http://www.skosmos.skos/test-concept-schemes/', $this->model);
+
+      $actual = $sparql->queryConceptSchemes('en');
+      $expected = array(
+          'http://exemple.fr/domains' => array(
+              'prefLabel' => 'Special Domains Concept Scheme'
+          ),
+          'http://exemple.fr/mt1' => array(
+              'prefLabel' => 'Micro-Thesaurus 1',
+              'subject' => array(
+                  'uri' => 'http://exemple.fr/d1',
+                  'prefLabel' => 'Domain 1'
+              )
+          ),
+          'http://exemple.fr/mt2' => array(
+              'prefLabel' => 'Micro-Thesaurus 2',
+              'subject' => array(
+                  'uri' => 'http://exemple.fr/d1',
+                  'prefLabel' => 'Domain 1'
+              )
+          ),
+          'http://exemple.fr/mt3' => array(
+              'prefLabel' => 'Micro-Thesaurus 3',
+              'subject' => array(
+                  'uri' => 'http://exemple.fr/d2',
+                  'prefLabel' => 'Domain 2'
+              )
+          ),
+          'http://exemple.fr/thesaurus' => array(
+              'prefLabel' => 'The Thesaurus'
+          ),
+      );
+
+      $this->assertEquals($expected, $actual);
   }
 
   /**
@@ -367,7 +552,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $voc2 = $this->model->getVocabulary('groups');
     $this->params->method('getSearchTerm')->will($this->returnValue('Carp'));
     $this->params->method('getVocabs')->will($this->returnValue(array($voc, $voc2)));
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', '?graph', $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', '?graph', $this->model);
     $actual = $sparql->queryConcepts(array($voc, $voc2), null, null, $this->params);
     $this->assertEquals(2, sizeof($actual));
     $this->assertEquals('http://www.skosmos.skos/groups/ta112', $actual[0]['uri']);
@@ -390,14 +575,14 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       // returns 3 concepts without the scheme limit, and only 2 with the scheme limit below
       $this->params->method('getSearchTerm')->will($this->returnValue('concept*'));
       $this->params->method('getSchemeLimit')->will($this->returnValue(array('http://www.skosmos.skos/multiple-schemes/cs1', 'http://www.skosmos.skos/multiple-schemes/cs2')));
-      $sparql = new GenericSparql('http://localhost:3030/ds/sparql', 'http://www.skosmos.skos/multiple-schemes/', $this->model);
+      $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', 'http://www.skosmos.skos/multiple-schemes/', $this->model);
       $actual = $sparql->queryConcepts(array($voc), null, null, $this->params);
       $this->assertEquals(2, sizeof($actual));
       $this->assertEquals('http://www.skosmos.skos/multiple-schemes/c1-in-cs1', $actual[0]['uri']);
       $this->assertEquals('http://www.skosmos.skos/multiple-schemes/c2-in-cs2', $actual[1]['uri']);
   }
-  
-  
+
+
   /**
    * @covers GenericSparql::queryConcepts
    * @covers GenericSparql::generateConceptSearchQueryCondition
@@ -416,6 +601,27 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->queryConcepts(array($voc), null, null, $this->params);
     $this->assertEquals(1, sizeof($actual));
     $this->assertEquals('Bass', $actual[0]['prefLabel']);
+  }
+
+  /**
+   * @covers GenericSparql::queryConcepts
+   * @covers GenericSparql::generateConceptSearchQueryCondition
+   * @covers GenericSparql::generateConceptSearchQueryInner
+   * @covers GenericSparql::generateConceptSearchQuery
+   * @covers GenericSparql::formatFilterGraph
+   * @covers GenericSparql::transformConceptSearchResults
+   * @covers GenericSparql::transformConceptSearchResult
+   * @covers GenericSparql::shortenUri
+   */
+  public function testQueryConceptsMatchLanguage()
+  {
+    $voc = $this->model->getVocabulary('test');
+    $this->params->method('getSearchTerm')->will($this->returnValue('e*'));
+    $this->params->method('getVocabIds')->will($this->returnValue(array('test')));
+    $this->params->method('getSearchLang')->will($this->returnValue('en'));
+    $actual = $this->sparql->queryConcepts(array($voc), null, null, $this->params);
+    $this->assertEquals(1, sizeof($actual));
+    $this->assertEquals('Eel', $actual[0]['prefLabel']);
   }
 
   /**
@@ -465,26 +671,6 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    * @covers GenericSparql::generateConceptSearchQueryInner
    * @covers GenericSparql::generateConceptSearchQuery
    * @covers GenericSparql::formatFilterGraph
-   * @covers GenericSparql::transformConceptSearchResults
-   * @covers GenericSparql::transformConceptSearchResult
-   * @covers GenericSparql::shortenUri
-   */
-  public function testQueryConceptsMatchNotation()
-  {
-    $voc = $this->model->getVocabulary('test');
-    $this->params->method('getSearchTerm')->will($this->returnValue('665*'));
-    $this->params->method('getVocabIds')->will($this->returnValue(array('test')));
-    $actual = $this->sparql->queryConcepts(array($voc), null, null, $this->params);
-    $this->assertEquals(1, sizeof($actual));
-    $this->assertEquals('http://www.skosmos.skos/test/ta112', $actual[0]['uri']);
-  }
-
-  /**
-   * @covers GenericSparql::queryConcepts
-   * @covers GenericSparql::generateConceptSearchQueryCondition
-   * @covers GenericSparql::generateConceptSearchQueryInner
-   * @covers GenericSparql::generateConceptSearchQuery
-   * @covers GenericSparql::formatFilterGraph
    */
   public function testQueryConceptsAdditionalFields()
   {
@@ -515,7 +701,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, sizeof($actual));
     $this->assertEquals('Bass', $actual[0]['prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryConcepts
    * @covers GenericSparql::generateConceptSearchQueryCondition
@@ -534,7 +720,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     foreach($actual as $match)
       $this->assertContains('bass', $match['prefLabel'], '',true);
   }
-  
+
   /**
    * @covers GenericSparql::queryConcepts
    * @covers GenericSparql::generateConceptSearchQueryCondition
@@ -553,7 +739,24 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     foreach($actual as $match)
       $this->assertContains('bass', $match['prefLabel'], '',true);
   }
-  
+
+  /**
+   * @covers GenericSparql::queryConcepts
+   * @covers GenericSparql::generateConceptSearchQueryCondition
+   * @covers GenericSparql::generateConceptSearchQueryInner
+   * @covers GenericSparql::generateConceptSearchQuery
+   * @covers GenericSparql::transformConceptSearchResults
+   * @covers GenericSparql::transformConceptSearchResult
+   * @covers GenericSparql::shortenUri
+   */
+  public function testQueryConceptsDoubleQuotesTerm()
+  {
+    $voc = $this->model->getVocabulary('test');
+    $this->params->method('getSearchTerm')->will($this->returnValue('"'));
+    $actual = $this->sparql->queryConcepts(array($voc), null, null, $this->params);
+    $this->assertEquals(0, sizeof($actual));
+  }
+
   /**
    * @covers GenericSparql::queryLabel
    * @covers GenericSparql::generateLabelQuery
@@ -574,7 +777,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $expected = array('en' => 'Carp');
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryLabel
    * @covers GenericSparql::generateLabelQuery
@@ -597,7 +800,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $expected = array('http://www.skosmos.skos/test/ta1' => array('label' => 'Fish'));
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryProperty
    * @covers GenericSparql::transformPropertyQueryResults
@@ -609,7 +812,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $expected = array();
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryProperty
    * @covers GenericSparql::transformPropertyQueryResults
@@ -630,22 +833,22 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $actual = $this->sparql->queryTransitiveProperty('http://www.skosmos.skos/test/ta111', array('skos:broader'), 'en', '10');
     $expected = array(
-      'http://www.skosmos.skos/test/ta111' => 
+      'http://www.skosmos.skos/test/ta111' =>
         array(
           'label' => 'Tuna',
-          'direct' => 
+          'direct' =>
           array (
             0 => 'http://www.skosmos.skos/test/ta1',
           ),
         ),
-        'http://www.skosmos.skos/test/ta1' => 
+        'http://www.skosmos.skos/test/ta1' =>
         array (
           'label' => 'Fish',
         )
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryTransitiveProperty
    * @covers GenericSparql::generateTransitivePropertyQuery
@@ -655,22 +858,22 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $actual = $this->sparql->queryTransitiveProperty('http://www.skosmos.skos/test/ta122', array('skos:broader'), 'en', '10');
     $expected = array(
-      'http://www.skosmos.skos/test/ta122' => 
+      'http://www.skosmos.skos/test/ta122' =>
         array (
           'label' => 'Black sea bass',
-          'direct' => 
+          'direct' =>
           array (
             0 => 'http://www.skosmos.skos/test/ta116',
           ),
         ),
-        'http://www.skosmos.skos/test/ta1' => 
+        'http://www.skosmos.skos/test/ta1' =>
         array (
           'label' => 'Fish',
         ),
-        'http://www.skosmos.skos/test/ta116' => 
+        'http://www.skosmos.skos/test/ta116' =>
         array (
           'label' => 'Bass',
-          'direct' => 
+          'direct' =>
           array (
             0 => 'http://www.skosmos.skos/test/ta1',
           ),
@@ -695,7 +898,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     foreach ($expected as $uri)
       $this->assertArrayHasKey($uri, $actual_uris);
   }
-  
+
   /**
    * @covers GenericSparql::queryChildren
    * @covers GenericSparql::generateChildQuery
@@ -746,7 +949,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       'http://www.skosmos.skos/test/ta116' => array (
         'uri' => 'http://www.skosmos.skos/test/ta116',
         'prefLabel' => 'Bass',
-        'broader' => 
+        'broader' =>
         array (
           0 => 'http://www.skosmos.skos/test/ta1',
         ),
@@ -754,7 +957,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       'http://www.skosmos.skos/test/ta122' => array (
         'uri' => 'http://www.skosmos.skos/test/ta122',
         'prefLabel' => 'Black sea bass',
-        'broader' => 
+        'broader' =>
         array (
           0 => 'http://www.skosmos.skos/test/ta116',
         ),
@@ -762,7 +965,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $props = array (
       'uri' => 'http://www.skosmos.skos/test/ta1',
-      'top' => 'http://www.skosmos.skos/test/conceptscheme',
+      'tops' => array('http://www.skosmos.skos/test/conceptscheme'),
       'prefLabel' => 'Fish',
     );
     $narrowers = array (
@@ -833,7 +1036,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('groups');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
     $actual = $sparql->ListConceptGroups('http://www.w3.org/2004/02/skos/core#Collection', 'en', false);
     $expected = array (0 => array ('prefLabel' => 'Fish', 'uri' => 'http://www.skosmos.skos/groups/fish', 'hasMembers' => true, 'childGroups' => array('http://www.skosmos.skos/groups/sub')), 1 => array ('prefLabel' => 'Freshwater fish', 'uri' => 'http://www.skosmos.skos/groups/fresh', 'hasMembers' => true), 2 => array ('prefLabel' => 'Saltwater fish', 'uri' => 'http://www.skosmos.skos/groups/salt', 'hasMembers' => true),3 => array ('prefLabel' => 'Submarine-like fish', 'uri' => 'http://www.skosmos.skos/groups/sub', 'hasMembers' => true));
     $this->assertEquals($expected, $actual);
@@ -848,12 +1051,12 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('groups');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
     $actual = $sparql->ListConceptGroupContents('http://www.w3.org/2004/02/skos/core#Collection', 'http://www.skosmos.skos/groups/salt', 'en');
     $this->assertEquals('http://www.skosmos.skos/groups/ta113', $actual[0]['uri']);
     $this->assertEquals(1, sizeof($actual));
   }
-  
+
   /**
    * @covers GenericSparql::listConceptGroupContents
    * @covers GenericSparql::generateConceptGroupContentsQuery
@@ -863,7 +1066,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
       $voc = $this->model->getVocabulary('showDeprecated');
       $graph = $voc->getGraph();
-      $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+      $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
       $actual = $sparql->ListConceptGroupContents('http://www.w3.org/2004/02/skos/core#Collection', 'http://www.skosmos.skos/groups/salt', 'en', $voc->getConfig()->getShowDeprecated());
       $expected = array (
           0 => array (
@@ -894,14 +1097,31 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('changes');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
-    $actual = $sparql->queryChangeList('en', 0, 'dc11:created');
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
+    $actual = $sparql->queryChangeList('dc:created', 'en', 0, 10);
     $order = array();
     foreach($actual as $concept) {
         array_push($order, $concept['prefLabel']);
     }
     $this->assertEquals(4, sizeof($actual));
     $this->assertEquals(array('Fourth date', 'Hurr Durr', 'Second date', 'A date'), $order);
+  }
+
+  /**
+   * @covers GenericSparql::queryChangeList
+   * @covers GenericSparql::generateChangeListQuery
+   * @covers GenericSparql::transFormChangeListResults
+   */
+  public function testMalformedDates() {
+    $voc = $this->model->getVocabulary('test');
+    $graph = $voc->getGraph();
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
+    $result = $sparql->queryChangeList('dc:modified', 'en', 0, 10);
+    $uris = array();
+    foreach($result as $concept) {
+      $uris[] = $concept['uri'];
+    }
+    $this->assertNotContains('http://www.skosmos.skos/test/ta114', $uris);
   }
 
   /**
@@ -912,7 +1132,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('test');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
     $this->params->method('getSearchTerm')->will($this->returnValue('*'));
     $this->params->method('getTypeLimit')->will($this->returnValue(array('mads:Topic')));
     $actual = $this->sparql->queryConcepts(array($voc), null, true, $this->params);
@@ -945,5 +1165,74 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
 );
     $this->assertEquals(array('en' => 'Bass'), $actual[0]['prefLabels']);
     $this->assertEquals(array(0 => array('uri' => 'http://www.skosmos.skos/test/ta1')), $actual[0]['skos:broader']);
+  }
+
+  /**
+   * @covers GenericSparql::querySuperProperties
+   */
+  public function testQuerySuperProperties()
+  {
+      $this->sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', '?graph', $this->model);
+      $actual = $this->sparql->querySuperProperties('http://example.com/myns#property');
+      $this->assertEquals(1, sizeof($actual));
+      $expected = array('http://example.com/myns#superProperty');
+      $this->assertEquals($actual, $expected);
+  }
+
+  /**
+   * @covers GenericSparql::queryAllConceptLabels
+   * @covers GenericSparql::generateAllLabelsQuery
+   * @covers GenericSparql::generateFromClause
+   */
+  public function testQueryAllConceptLabels()
+  {
+      $voc = $this->model->getVocabulary('test');
+      $graph = $voc->getGraph();
+      $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
+
+      $actual = $sparql->queryAllConceptLabels('http://www.skosmos.skos/test/ta112', 'en');
+
+      $this->assertTrue(array_key_exists('prefLabel',$actual));
+      $this->assertEquals($actual['prefLabel'][0], "Carp");
+
+      $this->assertTrue(array_key_exists('altLabel',$actual));
+      $this->assertEquals($actual['altLabel'][0], "Golden crucian");
+
+      $this->assertFalse(array_key_exists('hiddenLabel',$actual));
+  }
+
+  /**
+   * @covers GenericSparql::queryAllConceptLabels
+   * @covers GenericSparql::generateAllLabelsQuery
+   * @covers GenericSparql::generateFromClause
+   */
+  public function testQueryAllConceptLabelsNonexistentConcept()
+  {
+      $voc = $this->model->getVocabulary('test');
+      $graph = $voc->getGraph();
+      $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
+
+      $actual = $sparql->queryAllConceptLabels('http://www.skosmos.skos/test/notfound', 'en');
+
+      $this->assertNull($actual);
+  }
+
+  /**
+   * @covers GenericSparql::queryAllConceptLabels
+   * @covers GenericSparql::generateAllLabelsQuery
+   * @covers GenericSparql::generateFromClause
+   */
+  public function testQueryAllConceptLabelsNoPrefLabel()
+  {
+      $voc = $this->model->getVocabulary('test');
+      $graph = $voc->getGraph();
+      $sparql = new GenericSparql('http://localhost:13030/skosmos-test/sparql', $graph, $this->model);
+
+      $actual = $sparql->queryAllConceptLabels('http://www.skosmos.skos/test/ta112', 'sv');
+
+      $this->assertTrue(is_array($actual));
+      $this->assertFalse(array_key_exists('prefLabel',$actual));
+      $this->assertFalse(array_key_exists('altLabel',$actual));
+      $this->assertFalse(array_key_exists('hiddenLabel',$actual));
   }
 }
