@@ -1,27 +1,29 @@
 <?php
 
-class VocabularyCategoryTest extends PHPUnit_Framework_TestCase
+class VocabularyCategoryTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  private $model;
   private $mockres;
 
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->mockres = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
     $this->mockres->method('localName')->will($this->returnValue('local name'));
   }
-  
+
   /**
    * @covers VocabularyCategory::__construct
    * @expectedException Exception
-   * @expectedExceptionMessage Invalid constructor parameter given to DataObject. 
+   * @expectedExceptionMessage Invalid constructor parameter given to DataObject.
    */
   public function testConstructorWithInvalidParameters() {
-    new VocabularyCategory('invalid', 'invalid');
+    $vcat = new VocabularyCategory('invalid', 'invalid');
+    $this->assertNotNull($vcat);
   }
-  
+
   /**
    * @covers VocabularyCategory::__construct
    */
@@ -29,7 +31,7 @@ class VocabularyCategoryTest extends PHPUnit_Framework_TestCase
     $cat = new VocabularyCategory($this->model, $this->mockres);
     $this->assertInstanceOf('VocabularyCategory', $cat);
   }
-  
+
   /**
    * @covers VocabularyCategory::getVocabularies
    */
@@ -40,7 +42,7 @@ class VocabularyCategoryTest extends PHPUnit_Framework_TestCase
       $this->assertInstanceOf('Vocabulary', $voc);
     }
   }
-  
+
   /**
    * @covers VocabularyCategory::getTitle
    */
@@ -48,7 +50,7 @@ class VocabularyCategoryTest extends PHPUnit_Framework_TestCase
     $cat = new VocabularyCategory($this->model, $this->mockres);
     $this->assertEquals('local name', $cat->getTitle());
   }
-  
+
   /**
    * @covers VocabularyCategory::getTitle
    */
