@@ -32,11 +32,10 @@ class Honeypot
     {
         // Encrypt the current time
         $honey_time_encrypted = $this->getEncryptedTime();
-        $html = '<div id="' . $honey_name . '_wrap" style="display:none;">' . "\r\n" .
-                    '<input name="' . $honey_name . '" type="text" value="" id="' . $honey_name . '"/>' . "\r\n" .
-                    '<input name="' . $honey_time . '" type="text" value="' . $honey_time_encrypted . '"/>' . "\r\n" .
+        return '<div id="' . $honey_name . '_wrap" style="display:none;">' . "\r\n" .
+               '<input name="' . $honey_name . '" type="text" value="" id="' . $honey_name . '"/>' . "\r\n" .
+               '<input name="' . $honey_time . '" type="text" value="' . $honey_time_encrypted . '"/>' . "\r\n" .
                 '</div>';
-        return $html;
     }
     /**
     * Validate honeypot is empty
@@ -63,7 +62,7 @@ class Honeypot
         if ($this->disabled) {
             return true;
         }
-        
+
         // Get the decrypted time
         $value = $this->decryptTime($value);
         // The current time should be greater than the time the form was built + the speed option
