@@ -11,23 +11,31 @@ class ConceptProperty
     private $super;
     /** stores the property label */
     private $label;
+    /** stores the property tooltip */
+    private $tooltip;
     /** stores the property values */
     private $values;
     /** flag whether the values are sorted, as we do lazy sorting */
     private $is_sorted;
+    private $sort_by_notation;
 
     /**
      * Label parameter seems to be optional in this phase.
      * @param string $prop property type eg. 'rdf:type'.
-     * @param string $label
+     * @param string $label property label
+     * @param string $tooltip property tooltip/description
+     * @param string $super URI of superproperty
+     * @param boolean $sort_by_notation whether to sort the property values by their notation code
      */
-    public function __construct($prop, $label, $super=null)
+    public function __construct($prop, $label, $tooltip=null, $super=null, $sort_by_notation=false)
     {
         $this->prop = $prop;
         $this->label = $label;
+        $this->tooltip = $tooltip;
         $this->values = array();
         $this->is_sorted = true;
         $this->super = $super;
+        $this->sort_by_notation = $sort_by_notation;
     }
 
     /**
@@ -45,7 +53,7 @@ class ConceptProperty
         }
 
         // if not, see if there was a label for the property in the graph
-        if ($this->label) {
+        if ($this->label !== null) {
             return $this->label;
         }
 
@@ -54,19 +62,39 @@ class ConceptProperty
     }
 
     /**
-     * Returns a gettext translation for the property tooltip.
+     * Returns an alphanumeric ID for the property, suitable for use as a CSS identifier.
+     */
+    public function getID()
+    {
+        return preg_replace('/[^A-Za-z0-9-]/', '_', $this->prop);
+    }
+
+    /**
+     * Returns text for the property tooltip.
      * @return string
      */
     public function getDescription()
     {
         $helpprop = $this->prop . "_help";
 
-        return gettext($helpprop); // can't use string constant, it'd be picked up by xgettext
+        // see if we have a translation with the help text
+        $help = gettext($helpprop);
+        if ($help != $helpprop) {
+            return $help;
+        }
+
+       // if not, see if there was a comment/definition for the property in the graph
+        if ($this->tooltip !== null) {
+            return $this->tooltip;
+        }
+
+        // when nothing is found, don't show the tooltip at all
+        return null;
     }
 
     /**
      * Returns an array of the property values.
-     * @return array containing ConceptPropertyValue objects.
+     * @return ConceptMappingPropertyValue[]
      */
     public function getValues()
     {
@@ -78,16 +106,41 @@ class ConceptProperty
 
     public function addValue($value)
     {
-        $this->values[$value->getNotation() . $value->getLabel() . $value->getUri()] = $value;
+        $this->values[ltrim($value->getNotation() . ' ') . $value->getLabel() . rtrim(' ' . $value->getUri())] = $value;
         $this->is_sorted = false;
     }
 
     private function sortValues()
     {
+        # TODO: sort by URI as last resort
+        # Note that getLabel() returns URIs in case of no label and may return a prefixed value which affects sorting
         if (!empty($this->values)) {
-            uksort($this->values, function($a, $b) {
-                return strnatcasecmp($a,$b);
-            });
+            if ($this->sort_by_notation) {
+                uasort($this->values, function($a, $b) {
+                    $anot = $a->getNotation();
+                    $bnot = $b->getNotation();
+                    if ($anot == null) {
+                        if ($bnot == null) {
+                            // assume that labels are unique
+                            return strcoll(strtolower($a->getLabel()), strtolower($b->getLabel()));
+                        }
+                        return 1;
+                    }
+                    else if ($bnot == null) {
+                        return -1;
+                    }
+                    else {
+                        // assume that notations are unique
+                        return strnatcasecmp($anot, $bnot);
+                    }
+                });
+            }
+            else {
+                uasort($this->values, function($a, $b) {
+                    // assume that labels are unique
+                    return strcoll(strtolower($a->getLabel()), strtolower($b->getLabel()));
+                });
+            }
         }
         $this->is_sorted = true;
     }
@@ -100,7 +153,7 @@ class ConceptProperty
     {
         return $this->prop;
     }
-    
+
     /**
      * Returns property supertype (?property skos:subPropertyOf ?super) as a string.
      * @return string eg. 'skos:hiddenLabel'.
