@@ -11,6 +11,8 @@ class ConceptPropertyValue extends VocabularyDataObject
     private $type;
     /** content language */
     private $clang;
+    /** whether the property value is external w.r.t. to the subject resource */
+    private $external;
 
     public function __construct($model, $vocab, $resource, $prop, $clang = '')
     {
@@ -18,13 +20,22 @@ class ConceptPropertyValue extends VocabularyDataObject
         $this->submembers = array();
         $this->type = $prop;
         $this->clang = $clang;
+        // check if the resource is external to the current vocabulary
+        $this->external = ($this->getLabel('', 'null') === null);
+        if ($this->external) {
+            // if we find the resource in another vocabulary, use it instead
+            $exvocab = $this->getExVocab();
+            if ($exvocab !== null) {
+                $this->vocab = $exvocab;
+            }
+        }
     }
 
     public function __toString()
     {
         $label = is_string($this->getLabel()) ? $this->getLabel() : $this->getLabel()->getValue();
         if ($this->vocab->getConfig()->sortByNotation()) {
-            $label = $this->getNotation() . $label;
+            $label = ltrim($this->getNotation() . ' ') . $label;
         }
 
         return $label;
@@ -35,10 +46,26 @@ class ConceptPropertyValue extends VocabularyDataObject
         return $this->getEnvLang();
     }
 
-    public function getLabel($lang = '')
+    public function getLabel($lang = '', $fallbackToUri = 'uri')
     {
         if ($this->clang) {
             $lang = $this->clang;
+        }
+        if ($this->vocab->getConfig()->getLanguageOrder($lang)) {
+            foreach ($this->vocab->getConfig()->getLanguageOrder($lang) as $fallback) {
+                if ($this->resource->label($fallback) !== null) {
+                    return $this->resource->label($fallback);
+                }
+                // We need to check all the labels in case one of them matches a subtag of the current language
+                if ($this->resource->allLiterals('skos:prefLabel')) {
+                    foreach($this->resource->allLiterals('skos:prefLabel') as $label) {
+                        // the label lang code is a subtag of the UI lang eg. en-GB - create a new literal with the main language
+                        if ($label !== null && strpos($label->getLang(), $fallback . '-') === 0) {
+                            return EasyRdf\Literal::create($label, $fallback);
+                        }
+                    }
+                }
+            }
         }
 
         if ($this->resource->label($lang) !== null) { // current language
@@ -52,8 +79,18 @@ class ConceptPropertyValue extends VocabularyDataObject
         } elseif ($this->resource->getLiteral('rdf:value') !== null) { // any language
             return $this->resource->getLiteral('rdf:value');
         }
-        $label = $this->resource->shorten() ? $this->resource->shorten() : $this->getUri();
-        return $label;
+
+        // see if we can find a label in another vocabulary known by the skosmos instance
+        $label = $this->getExternalLabel($this->vocab, $this->getUri(), $lang);
+        if ($label) {
+            return $label;
+        }
+
+        if ($fallbackToUri == 'uri') {
+            // return uri if no label is found
+            return $this->resource->shorten() ? $this->resource->shorten() : $this->getUri();
+        }
+        return null;
     }
 
     public function getType()
@@ -66,6 +103,11 @@ class ConceptPropertyValue extends VocabularyDataObject
         return $this->resource->getUri();
     }
 
+    public function getExVocab()
+    {
+        return $this->model->guessVocabularyFromURI($this->getUri(), $this->vocab->getId());
+    }
+
     public function getVocab()
     {
         return $this->vocab;
@@ -73,7 +115,7 @@ class ConceptPropertyValue extends VocabularyDataObject
 
     public function getVocabName()
     {
-        return $this->vocab->getTitle();
+        return $this->vocab->getShortName();
     }
 
     public function addSubMember($member, $lang = '')
@@ -101,8 +143,7 @@ class ConceptPropertyValue extends VocabularyDataObject
     }
 
     public function isExternal() {
-        $propertyUris = $this->resource->propertyUris();
-        return empty($propertyUris);
+        return $this->external;
     }
 
     public function getNotation()
@@ -116,15 +157,15 @@ class ConceptPropertyValue extends VocabularyDataObject
     public function isReified() {
         return (!$this->resource->label() && $this->resource->getLiteral('rdf:value'));
     }
-    
+
     public function getReifiedPropertyValues() {
         $ret = array();
         $props = $this->resource->propertyUris();
         foreach($props as $prop) {
             $prop = (EasyRdf\RdfNamespace::shorten($prop) !== null) ? EasyRdf\RdfNamespace::shorten($prop) : $prop;
             foreach ($this->resource->allLiterals($prop) as $val) {
-                if ($prop !== 'rdf:value' && $this->resource->get($prop)) { // shown elsewhere
-                    $ret[gettext($prop)] = new ConceptPropertyValueLiteral($this->resource->get($prop), $prop);
+                if ($prop !== 'rdf:value') { // shown elsewhere
+                    $ret[gettext($prop)] = new ConceptPropertyValueLiteral($this->model, $this->vocab, $this->resource, $val, $prop);
                 }
             }
             foreach ($this->resource->allResources($prop) as $val) {
